fix: eeprom: mvebu: fix FDT offset and size that cause read issue from EEPROM
- issue: cannot load FDT from EEPROM correctly
- This patch fix the offset of FDT file in EEPROM.
  The issue is caused by the forced type conversion
  (u8) to the FDT file offset in EEPROM, since after
  the HW information is enlarged to 256 bytes, the
  FDT file offset is bigger than 256, however u8 forced
  conversion only keep lowest byte.

Reviewed-on: http://vgitil04.il.marvell.com:8080/27645
Tested-by: Hanna Hawa <hannah@marvell.com>

+++ board/mvebu/common/cfg_eeprom.h
@@ -63,7 +63,7 @@ struct eeprom_struct {
 struct config_types_info {
 	enum mv_config_type_id config_id;
 	char name[30];
-	u8 byte_num;
+	u32 byte_num;
 	u32 byte_cnt;
 };
 
@@ -83,7 +83,7 @@ struct hw_info_data_struct {
 	char value[HW_INFO_MAX_VALUE_LEN];
 };
 
-#define offset_in_eeprom(a)		((u8)(offsetof(struct eeprom_struct, a)))
+#define offset_in_eeprom(a)		((u32)(offsetof(struct eeprom_struct, a)))
 #define get_default_fdt_config_id(boardid)	mapping_default_fdt[boardid & (BOARD_ID_INDEX_MASK - 1)]
 #define boardid_is_valid(boardid)      (((boardid) >= MARVELL_BOARD_ID_BASE) && ((boardid) < MV_MAX_MARVELL_BOARD_ID))\
 					   ? true : false
